Revert "Jingyu987 branch list completed tests"
