+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -10,12 +10,10 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.ListCompletedCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.PrescriptionListParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
-import seedu.address.model.PrescriptionList;
 import seedu.address.model.ReadOnlyPrescriptionList;
 import seedu.address.model.prescription.Prescription;
 import seedu.address.storage.Storage;
@@ -34,7 +32,6 @@ public class LogicManager implements Logic {
     private final Model model;
     private final Storage storage;
     private final PrescriptionListParser prescriptionListParser;
-    private boolean isDisplayingCompletedList;
 
     /**
      * Constructs a {@code LogicManager} with the given {@code Model} and {@code Storage}.
@@ -43,25 +40,14 @@ public class LogicManager implements Logic {
         this.model = model;
         this.storage = storage;
         prescriptionListParser = new PrescriptionListParser();
-        this.isDisplayingCompletedList = false;
     }
 
     @Override
-    public CommandResult execute(String commandText) throws CommandException, ParseException, IOException {
+    public CommandResult execute(String commandText) throws CommandException, ParseException {
         logger.info("----------------[USER COMMAND][" + commandText + "]");
 
         CommandResult commandResult;
         Command command = prescriptionListParser.parseCommand(commandText);
-
-        if (command instanceof ListCompletedCommand) {
-            // Handle the ListCompletedCommand
-            this.isDisplayingCompletedList = true;
-            checkAndMoveEndedPrescriptions();
-        } else {
-            // Handle other commands...
-            this.isDisplayingCompletedList = false;
-        }
-
         commandResult = command.execute(model);
 
         try {
@@ -76,21 +62,6 @@ public class LogicManager implements Logic {
         return commandResult;
     }
 
-    /**
-     * Deletes prescriptions that are past the end date and stores them in the completed prescription list.
-     */
-    private void checkAndMoveEndedPrescriptions() throws IOException {
-        PrescriptionList prescriptionListCopy = new PrescriptionList(model.getPrescriptionList());
-        for (Prescription prescription : prescriptionListCopy.getPrescriptionList()) {
-            if (prescription.isEnded()) {
-                model.deletePrescription(prescription);
-                model.addCompletedPrescription(prescription);
-            }
-        }
-        storage.savePrescriptionList(model.getPrescriptionList());
-        storage.saveCompletedPrescriptionList(model.getCompletedPrescriptionList());
-    }
-
     @Override
     public ReadOnlyPrescriptionList getPrescriptionList() {
         return model.getPrescriptionList();
@@ -100,15 +71,7 @@ public class LogicManager implements Logic {
     public ObservableList<Prescription> getFilteredPrescriptionList() {
         return model.getFilteredPrescriptionList();
     }
-    @Override
-    public ObservableList<Prescription> getFilteredCompletedPrescriptionList() {
-        return model.getFilteredCompletedPrescriptionList();
-    }
 
-    @Override
-    public boolean getIsDisplayingCompletedList() {
-        return this.isDisplayingCompletedList;
-    }
     @Override
     public Path getPrescriptionListFilePath() {
         return model.getPrescriptionListFilePath();
