+++ src/main/java/seedu/address/logic/parser/PrescriptionListParser.java
@@ -14,7 +14,6 @@ import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.ListCompletedCommand;
 import seedu.address.logic.commands.ListTodayCommand;
 import seedu.address.logic.commands.TakeCommand;
 import seedu.address.logic.commands.UntakeCommand;
@@ -77,8 +76,6 @@ public class PrescriptionListParser {
             return new ExitCommandParser().parse(arguments);
         case HelpCommand.COMMAND_WORD:
             return new HelpCommandParser().parse(arguments);
-        case ListCompletedCommand.COMMAND_WORD:
-            return new ListCompletedCommandParser().parse(arguments);
         default:
             logger.finer("This user input caused a ParseException: " + userInput);
             throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
