+++ src/main/java/seedu/address/ui/CommandBox.java
@@ -1,7 +1,5 @@
 package seedu.address.ui;
 
-import java.io.IOException;
-
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
@@ -46,7 +44,7 @@ public class CommandBox extends UiPart<Region> {
         try {
             commandExecutor.execute(commandText);
             commandTextField.setText("");
-        } catch (CommandException | ParseException | IOException e) {
+        } catch (CommandException | ParseException e) {
             setStyleToIndicateCommandFailure();
         }
     }
@@ -81,7 +79,7 @@ public class CommandBox extends UiPart<Region> {
          *
          * @see seedu.address.logic.Logic#execute(String)
          */
-        CommandResult execute(String commandText) throws CommandException, ParseException, IOException;
+        CommandResult execute(String commandText) throws CommandException, ParseException;
     }
 
 }
