+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -97,12 +97,6 @@ public class LogicManagerTest {
         assertThrows(UnsupportedOperationException.class, () -> logic.getFilteredPrescriptionList().remove(0));
     }
 
-    @Test
-    public void getFilteredCompletedPrescriptionList_modifyList_throwsUnsupportedOperationException() {
-        assertThrows(UnsupportedOperationException.class, () -> logic.getFilteredCompletedPrescriptionList()
-                .remove(0));
-    }
-
     /**
      * Executes the command and confirms that
      * - no exceptions are thrown <br>
@@ -111,7 +105,7 @@ public class LogicManagerTest {
      * @see #assertCommandFailure(String, Class, String, Model)
      */
     private void assertCommandSuccess(String inputCommand, String expectedMessage,
-        Model expectedModel) throws CommandException, ParseException, IOException {
+        Model expectedModel) throws CommandException, ParseException {
         CommandResult result = logic.execute(inputCommand);
         assertEquals(expectedMessage, result.getFeedbackToUser());
         assertEquals(expectedModel, model);
