+++ src/test/java/seedu/address/logic/commands/EditCommandTest.java
@@ -11,7 +11,6 @@ import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_PROPRANOLO
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.CommandTestUtil.showPrescriptionAtIndex;
-import static seedu.address.testutil.CompletedPrescriptions.getCompletedPrescriptionList;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PRESCRIPTION;
 import static seedu.address.testutil.TypicalIndexes.INDEX_SECOND_PRESCRIPTION;
 import static seedu.address.testutil.TypicalPrescriptions.getTypicalPrescriptionList;
@@ -30,8 +29,7 @@ import seedu.address.testutil.EditPrescriptionDescriptorBuilder;
 import seedu.address.testutil.PrescriptionBuilder;
 
 public class EditCommandTest {
-    private Model model = new ModelManager(getTypicalPrescriptionList(), getCompletedPrescriptionList(),
-            new UserPrefs());
+    private Model model = new ModelManager(getTypicalPrescriptionList(), getTypicalPrescriptionList(), new UserPrefs());
 
     @Test
     public void execute_allFieldsSpecifiedUnfilteredList_success() {
@@ -43,7 +41,7 @@ public class EditCommandTest {
                 Messages.format(editedPrescription));
 
         Model expectedModel = new ModelManager(new PrescriptionList(model.getPrescriptionList()),
-                new PrescriptionList(getCompletedPrescriptionList()), new UserPrefs());
+                new PrescriptionList(getTypicalPrescriptionList()), new UserPrefs());
         expectedModel.setPrescription(model.getFilteredPrescriptionList().get(0), editedPrescription);
 
         assertCommandSuccess(editCommand, model, expectedMessage, expectedModel);
@@ -67,7 +65,7 @@ public class EditCommandTest {
                 Messages.format(editedPrescription));
 
         Model expectedModel = new ModelManager(new PrescriptionList(model.getPrescriptionList()),
-                new PrescriptionList(getCompletedPrescriptionList()), new UserPrefs());
+                new PrescriptionList(getTypicalPrescriptionList()), new UserPrefs());
 
         expectedModel.setPrescription(lastPrescription, editedPrescription);
 
@@ -84,7 +82,7 @@ public class EditCommandTest {
                 Messages.format(editedPrescription));
 
         Model expectedModel = new ModelManager(new PrescriptionList(model.getPrescriptionList()),
-                new PrescriptionList(getCompletedPrescriptionList()), new UserPrefs());
+                new PrescriptionList(getTypicalPrescriptionList()), new UserPrefs());
 
         assertCommandSuccess(editCommand, model, expectedMessage, expectedModel);
     }
@@ -104,7 +102,7 @@ public class EditCommandTest {
                 Messages.format(editedPrescription));
 
         Model expectedModel = new ModelManager(new PrescriptionList(model.getPrescriptionList()),
-                new PrescriptionList(getCompletedPrescriptionList()), new UserPrefs());
+                new PrescriptionList(getTypicalPrescriptionList()), new UserPrefs());
         expectedModel.setPrescription(model.getFilteredPrescriptionList().get(0), editedPrescription);
 
         assertCommandSuccess(editCommand, model, expectedMessage, expectedModel);
