+++ src/test/java/seedu/address/logic/commands/ExitCommandTest.java
@@ -1,7 +1,5 @@
 package seedu.address.logic.commands;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.ExitCommand.MESSAGE_EXIT_ACKNOWLEDGEMENT;
 
@@ -19,26 +17,4 @@ public class ExitCommandTest {
         CommandResult expectedCommandResult = new CommandResult(MESSAGE_EXIT_ACKNOWLEDGEMENT, false, true);
         assertCommandSuccess(new ExitCommand(), model, expectedCommandResult, expectedModel);
     }
-
-    @Test
-    public void equals() {
-        ExitCommand firstExitCommand =
-                new ExitCommand();
-        ExitCommand secondExitCommand =
-                new ExitCommand();
-
-        // same object -> returns true
-        assertTrue(firstExitCommand.equals(secondExitCommand));
-
-        // different types -> returns false
-        assertFalse(firstExitCommand.equals(1));
-        assertFalse(firstExitCommand.equals(true));
-        assertFalse(firstExitCommand.equals(new HelpCommand()));
-
-
-        // null -> returns false
-        assertFalse(firstExitCommand.equals(null));
-    }
-
-
 }
